+++ CotEditor/Sources/CELineNumberView.m
@@ -33,8 +33,10 @@
 
 
 static const CGFloat kMinVerticalThickness = 32.0;
+static const CGFloat kHorizontalThickness = 20.0;
 static const NSUInteger kMinNumberOfDigits = 3;
-static CGFloat const kLineNumberPadding = 3.0;
+static const CGFloat  kLineNumberPadding = 3.0;
+static const CGFloat kTicksLength = 4.0;
 
 
 @interface CELineNumberView ()
@@ -89,7 +91,8 @@ static const NSString *LineNumberFontName;
 {
     [super viewDidMoveToSuperview];
     
-    [self setRuleThickness:kMinVerticalThickness];
+    CGFloat thickness = [self orientation] == NSHorizontalRuler ? kHorizontalThickness : kMinVerticalThickness;
+    [self setRuleThickness:thickness];
 }
 
 
@@ -165,7 +168,7 @@ static const NSString *LineNumberFontName;
     CGFloat charWidth = advance.width;
     
     // prepare frame width
-    CGFloat width = [self ruleThickness];
+    CGFloat ruleThickness = [self ruleThickness];
     
     // adjust drawing coordinate
     NSPoint relativePoint = [self convertPoint:NSZeroPoint fromView:[self textView]];
@@ -179,12 +182,15 @@ static const NSString *LineNumberFontName;
     CFRelease(font);
     
     // add enough buffer to avoid broken drawing on Mountain Lion (10.8) with scroller (2015-07)
-    NSRect visibleRect = [[self textView] visibleRect];
+    NSRect visibleRect = [[self scrollView] documentVisibleRect];
     visibleRect.size.height += fontSize;
     
     // get glyph range which line number should be drawn
     NSRange visibleGlyphRange = [layoutManager glyphRangeForBoundingRect:visibleRect
                                                          inTextContainer:[[self textView] textContainer]];
+    
+    BOOL isVerticalText = [self orientation] == NSHorizontalRuler;
+    NSUInteger tailGlyphIndex = [layoutManager glyphIndexForCharacterAtIndex:[string length]];
     
     // counters
     NSUInteger glyphCount = visibleGlyphRange.location;
@@ -195,6 +201,46 @@ static const NSString *LineNumberFontName;
     NSRegularExpression *regex = [NSRegularExpression regularExpressionWithPattern:@"\n" options:0 error:nil];
     lineNumber += [regex numberOfMatchesInString:string options:0
                                         range:NSMakeRange(0, [layoutManager characterIndexForGlyphAtIndex:visibleGlyphRange.location])];
+    
+    // draw line number block
+    CGGlyph *digitGlyphsPtr = digitGlyphs;
+    CGFloat width = NSWidth([self bounds]);
+    void (^draw_number)(NSUInteger, NSUInteger, CGFloat, BOOL) = ^(NSUInteger lineNumber, NSUInteger lastLineNumber, CGFloat y, BOOL drawsNumber){
+        // translate x position for vertical text
+        if (isVerticalText) {
+            y = width + (visibleRect.origin.y + y) - fontSize / 2;
+        }
+        // draw line number
+        if (drawsNumber) {
+            NSUInteger digit = numberOfDigits(lineNumber);
+            
+            // calculate base position
+            CGPoint position;
+            if (isVerticalText) {
+                position = CGPointMake(y + (charWidth * (digit - 1)) / 2, ruleThickness + kTicksLength - 2);
+            } else {
+                position = CGPointMake(ruleThickness, y);
+            }
+            
+            // get glyphs and positions
+            CGGlyph glyphs[digit];
+            CGPoint positions[digit];
+            for (NSUInteger i = 0; i < digit; i++) {
+                position.x -= charWidth;
+                
+                positions[i] = position;
+                glyphs[i] = digitGlyphsPtr[numberAt(i, lineNumber)];
+            }
+            
+            CGContextShowGlyphsAtPositions(context, glyphs, positions, digit);  // draw line number
+        }
+        
+        if (isVerticalText) {
+            CGFloat x = floor(y - fontSize / 2) -0.5;
+            CGContextMoveToPoint(context, x, ruleThickness);
+            CGContextAddLineToPoint(context, x, ruleThickness - kTicksLength);
+        }
+    };
     
     // draw visible line numbers
     for (NSUInteger glyphIndex = visibleGlyphRange.location; glyphIndex < NSMaxRange(visibleGlyphRange); lineNumber++) { // count "real" lines
@@ -208,25 +254,29 @@ static const NSString *LineNumberFontName;
             CGFloat y = -NSMinY(lineRect);
             
             if (lastLineNumber == lineNumber) {  // wrapped line
-                CGPoint position = CGPointMake(width - charWidth, y);
-                CGContextShowGlyphsAtPositions(context, &wrappedMarkGlyph, &position, 1);  // draw wrapped mark
-                
-            } else {  // new line
-                NSUInteger digit = numberOfDigits(lineNumber);
-                
-                // get glyphs and positions
-                CGGlyph glyphs[digit];
-                CGPoint positions[digit];
-                for (NSUInteger i = 0; i < digit; i++) {
-                    glyphs[i] = digitGlyphs[numberAt(i, lineNumber)];
-                    positions[i] = CGPointMake(width - (i + 1) * charWidth, y);
+                if (!isVerticalText) {
+                    CGPoint position = CGPointMake(ruleThickness - charWidth, y);
+                    CGContextShowGlyphsAtPositions(context, &wrappedMarkGlyph, &position, 1);  // draw wrapped mark
                 }
                 
-                CGContextShowGlyphsAtPositions(context, glyphs, positions, digit);  // draw line number
+            } else {  // new line
+                BOOL drawsNumber = (!isVerticalText || lineNumber % 5 == 0 || lineNumber == 1);
+                draw_number(lineNumber, lastLineNumber, y, drawsNumber);
+            }
+            
+            glyphCount = NSMaxRange(range);
+            
+            // draw last line number anyway
+            if (isVerticalText &&  // vertical text
+                lastLineNumber != lineNumber &&  // new line
+                isVerticalText && lineNumber != 1 && lineNumber % 5 != 0 &&  // not yet drawn
+                tailGlyphIndex == glyphIndex &&  // last line
+                ![layoutManager extraLineFragmentTextContainer])  // no extra number
+            {
+                draw_number(lineNumber, lastLineNumber, y, YES);
             }
             
             lastLineNumber = lineNumber;
-            glyphCount = NSMaxRange(range);
         }
     }
     
@@ -235,25 +285,18 @@ static const NSString *LineNumberFontName;
         NSRect lineRect = [layoutManager extraLineFragmentUsedRect];
         CGFloat y = -NSMinY(lineRect);
         
-        NSUInteger digit = numberOfDigits(lineNumber);
-        
-        // get glyphs and positions
-        CGGlyph glyphs[digit];
-        CGPoint positions[digit];
-        for (NSUInteger i = 0; i < digit; i++) {
-            glyphs[i] = digitGlyphs[numberAt(i, lineNumber)];
-            positions[i] = CGPointMake(width - (i + 1) * charWidth, y);
-        }
-        
-        CGContextShowGlyphsAtPositions(context, glyphs, positions, digit);  // draw line number
+        draw_number(lineNumber, lastLineNumber, y, YES);
     }
     
+    CGContextStrokePath(context);  // draw vertical text tics
     CGContextRestoreGState(context);
     
     // adjust thickness
-    NSUInteger length = MAX(numberOfDigits(lineNumber), kMinNumberOfDigits);
-    CGFloat requiredWidth = MAX(length * charWidth + 3 * kLineNumberPadding, kMinVerticalThickness);
-    [self setRuleThickness:ceil(requiredWidth)];
+    if (!isVerticalText) {
+        NSUInteger length = MAX(numberOfDigits(lineNumber), kMinNumberOfDigits);
+        CGFloat requiredWidth = MAX(length * charWidth + 3 * kLineNumberPadding, kMinVerticalThickness);
+        [self setRuleThickness:ceil(requiredWidth)];
+    }
 }
 
 
@@ -271,6 +314,9 @@ static const NSString *LineNumberFontName;
 - (CGFloat)requiredThickness
 // ------------------------------------------------------
 {
+    if ([self orientation] == NSHorizontalRuler) {
+        return [self ruleThickness];
+    }
     return MAX(kMinVerticalThickness, [self ruleThickness]);
 }